Overview of the functions in the prospectr package

Misc functions for spectral data

This package implements a number of functions useful for pre-processing spectral data well as for selecting representative samples/spectra. The functions included here are particularly useful in Near-Infrared and Infrared Spectroscopy applications.

Details

This is the version 0.2.9 – proxy of the package. The main functionality is listed here.

Currently, the following preprocessing functions are available:

resample

resample2

movav

standardNormalVariate

msc

detrend

baseline

blockScale

blockNorm

binning

savitzkyGolay

gapDer

continuumRemoval

For the selection of representative samples/observations for calibrating spectral models the following functions ca be used:

naes

honigs

shenkWest

kenStone

duplex

puchwein

Other useful functions are also available:

read_nircal

readASD

spliceCorrection

cochranTest

See also

Author

Maintainer: Leonardo Ramirez-Lopez ramirez.lopez.leo@gmail.com

Authors:

Antoine Stevens (ORCID)

Leonardo Ramirez-Lopez (ORCID)