Import BUCHI NIRCal files

This function imports .nir files generated by BUCHI NIRCal software.

Usage

read_nircal(file, response = TRUE, spectra = TRUE,
            metadata = TRUE, progress = TRUE, verbose = TRUE)

Arguments

file

the name of the NIRCal (.nir) file which the data are to be read from. For URLs a temporary file is first downloaded and is then read.

response

a logical indicating if the data of the response variables must be returned (default is TRUE).

spectra

a logical indicating if the spectral data must be returned (default is TRUE).

metadata

a logical indicating if the metadada must be returned (default is TRUE).

progress

a logical indicating if a progress bar must be printed (default is TRUE).

verbose

a logical indicating if the number of spectra and response variables (an also the ID's of the spectra without gain and/or temperature information) must be printed (default is TRUE).

Value

a data.frame containing the metadata, response variables (if response = TRUE) and spectra (if spectra = TRUE, embedded in the data.frame as a matrix named ...$spc).

Details

The extension of the BUCHI NIRCal files is .nir. These files are used to store spectra generated by BUCHI N-500 and BUCHI NIRMaster FT-NIR sensors. See NIRCal technical data sheet.

Author

Leonardo Ramirez-Lopez